No uses cout

+++ Vector/goal.hpp
@@ -3,11 +3,11 @@ class cGoal{
 public:
 
 	float m_x, m_y;
-	int m_r, m_s, m_alpha;
+	float m_r, m_s, m_alpha;
 
 	cGoal(){};
 
-	cGoal(float x, float y, int r, int s, int alpha){
+	cGoal(float x, float y, float r, float s, float alpha){
 
 		m_x=x;
 		m_y=y;
